Do not run legacy input simulation tests in batch mode
